Fix overlap with grab window popping minimized windows back out
Also fixed being unable to minimize after reopening window from minimize. It appears that tkinter has a bug where if you change the alpha of a window after setting overrideredirect, it will reset the overrideredirect state of the window. Opted not to go with a solution where you reset overrideredirect after resetting the alpha because it was causing black artifacts to appear when reloading the window.

Currently, there are transparent effects when reloading window (it seems that there is some visual lag). It was easy to avoid having visual artifacts on the window when it came back out if I used alpha to control visibility, but then the above issues with overrideredirect would appear and my window showed on the task bar when I didn't want it to, so that solution has been scrapped for now.